Converge jcs & rdfc algorithms into 1 suite.
