+++ tests/80-algorithms.js
@@ -1,0 +1,11 @@
+/*!
+ * Copyright (c) 2022-2024 Digital Bazaar, Inc.
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+import {
+  jcs,
+  rdfc
+} from './suites/algorithms.js';
+
+jcs();
+rdfc();
